Moved the base menu pieces into their own folder in pieces
