+++ src/pieces/discussion/menu/staff/ViewStudentsMenu.ts
@@ -5,8 +5,8 @@ import { sendDismissableReply } from "../../../../generalUtilities/DismissableMe
 import { getCourseByName } from "../../../../generalUtilities/getCourseByName";
 import { GUILDS } from "../../../../secret";
 import { addScorePeriods } from "../../tracking/scoreFunctions";
-import { ComponentBehavior } from "../BaseMenu";
-import { CustomNavOptions, NavigatedMenu, NavigatedMenuData } from "../NavigatedMenu";
+import { ComponentBehavior } from "../../../menu/BaseMenu";
+import { CustomNavOptions, NavigatedMenu, NavigatedMenuData } from "../../../menu/NavigatedMenu";
 import { updateToManageCourseMenu } from "./ManageCourseMenu";
 
 const TITLE_COURSE_PREFIX = "Students of CISC ";
