+++ src/pieces/discussion/menu/staff/manageScorePeriods/PeriodModalUtilities.ts
@@ -3,7 +3,7 @@ import { DateTime } from "luxon";
 import { courseModel } from "../../../../../generalModels/Course";
 import { ScorePeriod } from "../../../../../generalModels/DiscussionScoring";
 import { scoreAllThreads } from "../../../tracking/scoreFunctions";
-import { ModalInputHandler, createDiscussionModal } from "../../ModalUtilities";
+import { ModalInputHandler, createDiscussionModal } from "../../../../menu/ModalUtilities";
 import { refreshManagePeriodsMenu, updateToManagePeriodsMenu } from "./ManageScorePeriodsMenu";
 import { DATABASE_ERROR_MESSAGE, DATE_STRING_FORMAT, END_DATE_INPUT_ID, GOAL_POINTS_INPUT_ID, INVALID_END_DATE_REASON, INVALID_GOAL_POINTS_REASON, INVALID_INDEX_PERIOD_REASON, INVALID_MAX_POINTS_REASON, INVALID_START_DATE_REASON, MAX_POINTS_INPUT_ID, START_DATE_INPUT_ID } from "./ModalComponents";
 import { getCourseByName } from "../../../../../generalUtilities/getCourseByName";
